+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * changed level of update token log record from info to debug
+* created refresh token ratio parameter
 
 ## 1.9.0